requirements.txt
discord.py
requests
google-generativeai
Pillow